% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{compare_matrices}
\alias{compare_matrices}
\title{Compare deux matrices}
\usage{
compare_matrices(mat1, mat2)
}
\arguments{
\item{mat1}{Première matrice à comparer}

\item{mat2}{Deuxième matrice à comparer}
}
\value{
Matrice de comparaison
}
\description{
Cette fonction compare deux matrices et retourne une matrice de comparaison.
}
