% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrices.R
\name{compteur}
\alias{compteur}
\title{compteur calcule le nombre de 1 successifs pour chaque ligne ou chaque colonne d'une matrice carrée comportant des 1 et des 0}
\usage{
compteur(M, t)
}
\arguments{
\item{M}{une matrice carrée remplie de 0 et 1}

\item{t}{un nombre prenant la valeur 1 pour compter par colonne}
}
\description{
compteur calcule le nombre de 1 successifs pour chaque ligne ou chaque colonne d'une matrice carrée comportant des 1 et des 0
}
\examples{
compteur(matrix(c(0,0,1,1,1,1,0,1,1,1,0,1,1,0,0,0),nrow=4,ncol=4,byrow=TRUE),0)
}
\author{
Mottier
}
