% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff.R
\name{difficulte}
\alias{difficulte}
\title{difficulte}
\usage{
difficulte(s)
}
\arguments{
\item{s}{chaîne de caractères parmi "Facile", "Moyen", "Difficile"}
}
\description{
difficulte convertit le niveau de difficulté choisi par l'utilisateur en une proportion p de cases noires : 70\% pour le mode "Facile", 
55\% pour le mode "Moyen" et 45\% pour le mode "Diificile".
}
\examples{
difficulte("Facile")
}
\author{
Bouland - Mottier
}
