% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{generer_grille_aleatoire}
\alias{generer_grille_aleatoire}
\title{Jeu de Picross Shiny}
\usage{
generer_grille_aleatoire(taille, p)
}
\description{
Cet ensemble de fonctions et d'interface Shiny permet de jouer au jeu de Picross.
}
