% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matrices.R
\name{matrice_alea}
\alias{matrice_alea}
\title{matrice_alea}
\usage{
matrice_alea(n, p)
}
\arguments{
\item{n}{un entier}

\item{p}{un réel de [0,1]}
}
\description{
matrice_alea crée une matrice carrée aléatoire de taille n contenant des 0 et 1, avec une proportion p de 1.
}
\examples{
matrice_alea(5,0.6)
}
\author{
Bouland - Mottier
}
