% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{obtenir_indices_colonne}
\alias{obtenir_indices_colonne}
\title{Obtient les indices de colonne contigus}
\usage{
obtenir_indices_colonne(colonne)
}
\arguments{
\item{colonne}{Vecteur de la colonne à analyser}
}
\value{
Indices des colonnes contigus
}
\description{
Cette fonction obtient les indices des colonnes contigus dans une matrice.
}
