% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{obtenir_indices_ligne}
\alias{obtenir_indices_ligne}
\title{Obtient les indices de ligne contigus}
\usage{
obtenir_indices_ligne(ligne)
}
\arguments{
\item{ligne}{Vecteur de la ligne à analyser}
}
\value{
Indices des lignes contigus
}
\description{
Cette fonction obtient les indices des lignes contigus dans une matrice.
}
