% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{server}
\alias{server}
\title{server}
\usage{
server(input, output)
}
\description{
Logique serveur pour le jeu de Picross

Cette fonction définit la logique serveur pour le jeu de Picross en utilisant Shiny.
}
