% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\docType{data}
\name{ui}
\alias{ui}
\title{UI}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
UI pour le jeu de Picross

Cette fonction définit l'interface utilisateur (UI) pour le jeu de Picross en utilisant Shiny.
}
\keyword{datasets}
