% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{update_player_matrix}
\alias{update_player_matrix}
\title{Mets à jour la matrice du joueur}
\usage{
update_player_matrix(i, j)
}
\arguments{
\item{i}{indice ligne.}

\item{j}{indice colonne.}
}
\value{
A matrix.
}
\description{
Mets à jour la matrice du joueur
}
\examples{
update_player_matrix(1, 1)
update_player_matrix(3, 4)
}
