% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picross.R
\name{verifier_positions}
\alias{verifier_positions}
\title{Vérifie la position de 1 dans deux matrices}
\usage{
verifier_positions(matrice1, matrice2)
}
\arguments{
\item{matrice1}{: première matrice}

\item{matrice2}{: deuxième matrice}
}
\value{
TRUE or FALSE
}
\description{
Vérifie la position de 1 dans deux matrices
}
\examples{
verifier_positions(matrix(c(1,0,1,1), nrow = 2 , ncol=2),matrix(c(1,0,1,1), nrow = 2 , ncol=2))
verifier_positions(matrix(c(1,0,1,1,1,0), nrow = 2 , ncol=3),matrix(c(1,0,1,0,1,0), nrow = 2 , ncol=3))
}
