% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diff.R
\name{difficulte}
\alias{difficulte}
\title{difficulte convertit le niveau de difficulté choisi par l'utilisateur en une proportion p de cases noires. 
Easy -> 0.8
Medium -> 0.7
Hard -> 0.6}
\usage{
difficulte(s)
}
\arguments{
\item{s}{: chaîne de caractères parmi "Easy", "Medium", "Hard"}
}
\description{
difficulte convertit le niveau de difficulté choisi par l'utilisateur en une proportion p de cases noires. 
Easy -> 0.8
Medium -> 0.7
Hard -> 0.6
}
\examples{
difficulte("Easy")
}
\author{
Bouland - Mottier
}
